requirements.txt
Django==1.11.11
django-oauth-toolkit==1.1.2
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
oauthlib==2.1.0
paramiko==2.4.1
salt==2018.3.0
coreapi==2.3.3
